pyproject.toml
[tool.poetry]
name = "agentopia"
version = "0.1.0"
description = ""
authors = ["Yash Agarwal <heisenberg@buffer.finance>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.3.3"
requests = "^2.32.3"
eth-account = "^0.13.4"
web3 = "^7.5.0"
pysha3 = "^1.0.2"
pipe = "^2.2"
pydantic-settings = "^2.6.1"
pytz = "^2024.2"
python-dateutil = "^2.9.0.post0"
orjson = "^3.10.11"
fastapi = "^0.115.5"
openai = "^1.55.0"
mkdocs = "^1.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
